pyproject.toml
[tool.poetry]
name = "transactions-analysis"
version = "0.1.0"
description = ""
authors = ["Arctiqa <sinitro@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.3"
xlrd = "^2.0.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
openpyxl = "^3.1.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
mypy = "^1.7.1"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
